rename: 페이지 파일 정리
