+++ src/pages/listPage/GoodsListPage.jsx
@@ -1,12 +1,12 @@
 import React from "react";
 import styled from "styled-components";
-import Header from "../components/common/Header";
-import Footer from "../components/common/Footer";
-import Sort from "../components/ListPage/Sort";
-import Pagination from "../components/ListPage/Pagination";
-import List from "../components/ListPage/List";
-import { ListPage, ListItem } from "../components/ListPage/ListStyle";
-import dummy from "../components/ListPage/dummy";
+import Header from "../../components/common/Header";
+import Footer from "../../components/common/Footer";
+import Sort from "../../components/ListPage/Sort";
+import Pagination from "../../components/ListPage/Pagination";
+import List from "../../components/ListPage/List";
+import { ListPage, ListItem } from "../../components/ListPage/ListStyle";
+import dummy from "../../components/ListPage/dummy";
 
 const GoodsListPage = () => {
   return (
